+++ GbsoDev.TechTest.Library.Dal/LibroDal.cs
@@ -4,14 +4,14 @@ using Microsoft.EntityFrameworkCore;
 
 namespace GbsoDev.TechTest.Library.Dal
 {
-	internal sealed class LibroDal : BaseDataAccesLayer<Libro, int>, ILibroDal
+	internal sealed class LibroDal : BaseDataAccesLayer<Libro, long>, ILibroDal
 	{	
 		public LibroDal(IServiceProvider serviceProvider) : base(serviceProvider)
 		{
 
 		}
 
-		public List<Libro> ListLibrosByAutorId(int id)
+		public List<Libro> ListLibrosByAutorId(long id)
 		{
 			return RootContext.Autores
 				.Include(x => x.Libros)
